Big update for supporting Modbus master
[*] Update s2files.xml with IDs
[*] Update AVM-KIV, AVM-KDV, AVM-KTF config
[*] Moving 'ModbusItem' data structure to S2 lib
[*] Update 'WidgetFactory' class for using later documentation
[*] Small changes

+++ src/s2/include/s2/modbusitem.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <interfaces/types/modbus_types.h>
+#include <QtCore/QtGlobal>
 
 namespace ModbusItem
 {
@@ -82,7 +82,7 @@ struct Item
     TypeId typedata;    ///< Тип данных
     quint8 func;        ///< код функции modbus
     quint16 reg;        ///< Начальный адрес регистра
-    quint8 cnt;         ///< Количество регистров
+    quint16 cnt;        ///< Количество регистров
 
     friend bool operator==(const Item &lhs, const Item &rhs);
     friend bool operator!=(const Item &lhs, const Item &rhs);
